+++ src/main/java/ca/bc/gov/api/oracle/legacy/exception/NoSearchParameterFound.java
@@ -1,12 +1,12 @@
-package ca.bc.gov.api.oracle.legacy.exceptions;
+package ca.bc.gov.api.oracle.legacy.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.server.ResponseStatusException;
 
 @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class InvalidClientNumberException extends ResponseStatusException {
-  public InvalidClientNumberException() {
-    super(HttpStatus.BAD_REQUEST, "Invalid Client Number");
+public class NoSearchParameterFound extends ResponseStatusException {
+  public NoSearchParameterFound() {
+    super(HttpStatus.BAD_REQUEST, "No search parameter found");
   }
 }
