+++ src/main/java/ca/bc/gov/api/oracle/legacy/service/ClientService.java
@@ -3,9 +3,9 @@ package ca.bc.gov.api.oracle.legacy.service;
 import ca.bc.gov.api.oracle.legacy.dto.ClientPublicViewDto;
 import ca.bc.gov.api.oracle.legacy.entity.ClientPublicViewEntity;
 import ca.bc.gov.api.oracle.legacy.entity.QClientPublicViewEntity;
-import ca.bc.gov.api.oracle.legacy.exceptions.ClientNotFoundException;
-import ca.bc.gov.api.oracle.legacy.exceptions.InvalidClientNumberException;
-import ca.bc.gov.api.oracle.legacy.exceptions.NoSearchParameterFound;
+import ca.bc.gov.api.oracle.legacy.exception.ClientNotFoundException;
+import ca.bc.gov.api.oracle.legacy.exception.InvalidClientNumberException;
+import ca.bc.gov.api.oracle.legacy.exception.NoSearchParameterFound;
 import ca.bc.gov.api.oracle.legacy.repository.ClientPublicViewRepository;
 import ca.bc.gov.api.oracle.legacy.util.ClientMapper;
 import com.querydsl.core.types.dsl.BooleanExpression;
